Quick and dirty remove align_start_time parameter added in Kairos 0.94
This makes two changes:
	- Default parameter in UI to false
	- Don't supply parameter unless set to true

Why: Kairos before 0.94 didn't support this

align_start_time was added to kairosdb in commit 1b163ae

	kairosdb/kairosdb@1b163aec153573dc5df4ddc2976d6ba2b58603bc

For the 0.94 release.

Ideally the proper fix would be a plugin configuration flag (or
autodetect) the Kairos version and add/remove parameters to API
calls and GUI based on that.